Fixed circular reference, log_generator now uses contexts
