+++ onix/tests/test_utilities.py
@@ -353,157 +353,3 @@ class TestDetermineHiddenPowerType(object):
     def test_weird_ivs(self):
         ivs = PokeStats(4, 21, 10, 20, 23, 11)
         assert 'grass' == utilities.determine_hidden_power_type(ivs)
-
-
-class TestGetAllFormes(object):
-
-    @classmethod
-    def setup_class(cls):
-        cls.context = contexts.get_standard_context()
-
-    def test_pokemon_with_a_single_forme(self):
-
-        expected = [Forme('stunfisk', 'static',
-                          PokeStats(109, 66, 84, 81, 99, 32))]
-        assert expected == utilities.get_all_formes('stunfisk', 'static', None,
-                                                   ['voltswitch'],
-                                                    self.context)
-
-    def test_pokemon_with_wrong_ability(self):
-
-        expected = [Forme('vileplume', 'chlorophyll',
-                          PokeStats(75, 80, 85, 110, 90, 50))]
-        assert expected == utilities.get_all_formes('vileplume', 'flashfire',
-                                                   'absorbbulb',
-                                                   ['gigadrain'],
-                                                    self.context)
-
-    def test_pokemon_with_wrong_ability_in_hackmons(self):
-
-        expected = [Forme('vileplume', 'flashfire',
-                          PokeStats(75, 80, 85, 110, 90, 50))]
-        assert expected == utilities.get_all_formes('vileplume', 'flashfire',
-                                                  'absorbbulb',
-                                                  ['gigadrain'],
-                                                    self.context,
-                                                  True, True)
-
-    def test_pokemon_with_wrong_ability_in_aaa(self):
-
-        expected = [Forme('vileplume', 'flashfire',
-                          PokeStats(75, 80, 85, 110, 90, 50))]
-        assert expected == utilities.get_all_formes('vileplume', 'flashfire',
-                                                  'absorbbulb',
-                                                  ['gigadrain'],
-                                                    self.context,
-                                                  False, True)
-
-    def test_pokemon_with_mega_forme(self):
-
-        expected = [Forme('venusaur', 'chlorophyll',
-                          PokeStats(80, 82, 83, 100, 100, 80)),
-                    Forme('venusaurmega', 'thickfat',
-                          PokeStats(80, 100, 123, 122, 120, 80))]
-
-        assert set(expected) == set(
-            utilities.get_all_formes('venusaur', 'chlorophyll',
-                                     'venusaurite', ['frenzyplant'],
-                                     self.context))
-        assert set(expected) == set(
-            utilities.get_all_formes('venusaurmega', 'chlorophyll',
-                                     'venusaurite', ['frenzyplant'],
-                                     self.context))
-
-    def test_pokemon_with_multiple_mega_formes(self):
-
-        expected = [Forme('charizard', 'blaze',
-                          PokeStats(78, 84, 78, 109, 85, 100)),
-                    Forme('charizardmegay', 'drought',
-                          PokeStats(78, 104, 78, 159, 115, 100))]
-
-        assert set(expected) == set(
-            utilities.get_all_formes('charizard', 'blaze', 'charizarditey',
-                                     ['blastburn'],
-                                     self.context))
-
-    def test_mega_mon_without_stone(self):
-
-        expected = [Forme('venusaur', 'chlorophyll',
-                          PokeStats(80, 82, 83, 100, 100, 80))]
-
-        assert expected == utilities.get_all_formes('venusaur', 'chlorophyll',
-                                                    'leftovers',
-                                                    ['frenzyplant'],
-                                                    self.context)
-
-        assert expected == utilities.get_all_formes('venusaurmega',
-                                                    'chlorophyll',
-                                                    'leftovers',
-                                                    ['frenzyplant'],
-                                                    self.context)
-
-    def test_mega_mon_without_stone_in_hackmons(self):
-
-        expected = [Forme('venusaurmega', 'chlorophyll',
-                          PokeStats(80, 100, 123, 122, 120, 80))]
-
-        assert expected == utilities.get_all_formes('venusaurmega',
-                                                    'chlorophyll',
-                                                    'leftovers',
-                                                    ['frenzyplant'],
-                                                    self.context,
-                                                    True, True)
-
-    def test_mega_evolving_mega_in_hackmons(self):
-
-        expected = [Forme('mewtwomegax', 'steadfast',
-                          PokeStats(106, 190, 100, 154, 100, 130)),
-                    Forme('mewtwomegay', 'marvelscale',
-                          PokeStats(106, 150, 70, 194, 120, 140))
-                    ]
-
-        assert set(expected) == set(
-            utilities.get_all_formes('mewtwomegay', 'marvelscale', 'mewtwonitex',
-                                     ['psychic'],
-                                     self.context, True, True))
-
-    def test_aegislash_blade_in_hackmons(self):
-
-        expected = [Forme('aegislash', 'stancechange',
-                          PokeStats(60, 50, 150, 50, 150, 60)),
-                    Forme('aegislashblade', 'stancechange',
-                          PokeStats(60, 150, 50, 150, 50, 60))]
-
-        assert set(expected) == set(
-            utilities.get_all_formes('aegislashblade', 'stancechange', None,
-                                     ['kingsshield'],
-                                     self.context, True, True))
-
-        expected = [Forme('aegislashblade', 'stancechange',
-                          PokeStats(60, 150, 50, 150, 50, 60))]
-
-        assert set(expected) == set(
-            utilities.get_all_formes('aegislashblade', 'stancechange', None,
-                                     ['shadowball'],
-                                     self.context, True, True))
-
-        expected = [Forme('aegislashblade', 'contrary',
-                          PokeStats(60, 150, 50, 150, 50, 60))]
-
-        assert set(expected) == set(
-            utilities.get_all_formes('aegislashblade', 'contrary', None,
-                                     ['kingsshield'],
-                                     self.context, True, True))
-
-    def test_unrecognized_condition_type(self):
-
-        self.context.accessible_formes['zapdos'] = [[{'because':
-                                                          'i feel like it'},
-                                              ['moltres']]]
-        utilities.get_all_formes('quilava', 'flashfire', None, ['flamewheel'],
-                                 self.context)
-
-        with pytest.raises(ValueError):
-            utilities.get_all_formes('zapdos', 'pressure', 'leftovers',
-                                     ['voltswitch'],
-                                     self.context)
